requirements.txt
flask
torch
torchvision
numpy
pillow
opencv-python
pandas
scikit-learn
efficientnet-pytorch
